jobs/hello-world/pyproject.toml
[tool.poetry]
name = "hello-world"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "hello_world"}]

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-secret-manager = "^2.15.1"
google-crc32c = "^1.5.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
